# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface DestinyRequest {
  name: string;
  birthDate: string;
  birthTime?: string;
  gender?: string;
  palmCategory?: string;
}

export interface DestinyResult {
  personalityProfile: string;
  strengths: string[];
  weaknesses: string[];
  hiddenTalents: string;
  loveStyle: string;
  moneyEnergy: string;
  futurePrediction: string;
  soulmateType: string;
  auraType: string;
  auraColor: string;
  destinyScore: number;
  luckyNumbers: number[];
  luckyDays: string[];
  zodiacSign: string;
  destinyArchetype: string;
  lifePathNumber: number;
}

export interface DailyFortuneRequest {
  name: string;
  zodiacSign: string;
  auraType?: string;
  lifePathNumber?: number;
  destinyArchetype?: string;
  birthDate?: string;
}

export interface DailyFortune {
  message: string;
  energyLevel: number;
  luckyColor: string;
  affirmation: string;
  warningArea: string;
  spiritAnimal: string;
  spiritAnimalEmoji: string;
  loveInsight: string;
  careerInsight: string;
  healthInsight: string;
  bestTime: string;
  luckyNumber: number;
  cosmicTip: string;
}

export interface CompatibilityRequest {
  userName: string;
  userBirthDate: string;
  userZodiacSign: string;
  partnerName: string;
  partnerBirthDate: string;
}

export interface CompatibilityResult {
  score: number;
  partnerZodiacSign: string;
  summary: string;
  loveCompatibility: number;
  friendshipCompatibility: number;
  passionLevel: number;
  strengths: string[];
  challenges: string[];
  advice: string;
  sharedDestiny: string;
}

export interface PublicUser {
  id: number;
  email: string | null;
  displayName: string | null;
  vkLinked: boolean;
  emailVerified: boolean;
  birthDate?: string | null;
  birthTime?: string | null;
  gender?: string | null;
}

export interface UpdateMeRequest {
  displayName?: string | null;
  birthDate?: string | null;
  birthTime?: string | null;
  gender?: string | null;
}

export interface AuthSession {
  token: string;
  user: PublicUser;
  mailerWarning?: string | null;
}

export interface AuthRegisterRequest {
  email: string;
  password: string;
  displayName?: string;
}

export interface AuthLoginRequest {
  email: string;
  password: string;
}

export interface MeResponse {
  user: PublicUser;
}

export interface AuthForgotPasswordRequest {
  email: string;
}

export interface ForgotPasswordResult {
  ok: boolean;
  mailerConfigured: boolean;
}

export interface AuthResetPasswordRequest {
  token: string;
  newPassword: string;
}

export interface VkConfigResponse {
  configured: boolean;
  authUrl?: string;
  redirectUri?: string;
  state?: string;
  error?: string;
}

export interface PersonalData {
  email: string | null;
  phone: string | null;
  city: string | null;
  marketingConsent: boolean;
  personalizationConsent: boolean;
  updatedAt: string | null;
}

export interface PersonalDataResponse {
  data: PersonalData;
}

export interface PersonalDataRequest {
  email?: string | null;
  phone?: string | null;
  city?: string | null;
  marketingConsent: boolean;
  personalizationConsent: boolean;
  policyVersion?: string;
}

export interface ConsentRequiredError {
  error: string;
  requiredPolicyVersion: string;
}

export interface ConsentRecord {
  id: number;
  policyVersion: string;
  marketingConsent: boolean;
  personalizationConsent: boolean;
  acceptedAt: string;
  ip?: string | null;
  userAgent?: string | null;
}

export interface ConsentHistoryResponse {
  currentPolicyVersion: string;
  records: ConsentRecord[];
}

export interface AcceptConsentRequest {
  policyVersion: string;
  marketingConsent: boolean;
  personalizationConsent: boolean;
}

export interface ConsentRecordResponse {
  record: ConsentRecord;
}

export type ExportResponseUser = { [key: string]: unknown };

export interface ExportResponse {
  exportedAt: string;
  policyVersion: string;
  user: ExportResponseUser;
  personalData?: PersonalData | null;
  consentRecords: ConsentRecord[];
}

export interface DeleteAccountRequest {
  confirm: string;
}

export interface VkExchangeRequest {
  code: string;
  redirectUri: string;
  state: string;
}

export type AuthVerifyEmailParams = {
token: string;
};

export type AuthVkCallbackParams = {
code?: string;
state?: string;
error?: string;
};