# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
...
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
    patch:
      operationId: authUpdateMe
      tags: [auth]
      summary: Update fields on the current user (first-login profile attach)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
  /auth/verify-email:
    get:
      operationId: authVerifyEmail
      tags: [auth]
      summary: Verify email via token from confirmation link (returns HTML)
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: HTML confirmation page
          content:
            text/html:
              schema:
                type: string
        "400":
          description: Invalid or expired link
          content:
            text/html:
              schema:
                type: string
  /auth/vk/callback:
    get:
      operationId: authVkCallback
      tags: [auth]
      summary: VK OAuth redirect target (returns HTML)
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Browser landing page
          content:
            text/html:
              schema:
                type: string
  /auth/forgot-password:
    post:
      operationId: authForgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthForgotPasswordRequest"
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResult"
  /auth/reset-password:
    post:
      operationId: authResetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /auth/vk/exchange:
    post:
      operationId: authVkExchange
      tags: [auth]
      summary: Exchange VK code for session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VkExchangeRequest"
      responses:
        "200":
          description: Logged in via VK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    PublicUser:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        vkLinked:
          type: boolean
        emailVerified:
          type: boolean
        birthDate:
          type: string
          nullable: true
        birthTime:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
      required:
        - id
        - email
        - displayName
        - vkLinked
        - emailVerified
    UpdateMeRequest:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        birthTime:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/PublicUser"
        mailerWarning:
          type: string
          nullable: true
      required:
        - token
        - user
    AuthRegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
      required:
        - email
        - password
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    MeResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/PublicUser"
      required:
        - user
    AuthForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    ForgotPasswordResult:
      type: object
      properties:
        ok:
          type: boolean
        mailerConfigured:
          type: boolean
      required:
        - ok
        - mailerConfigured
    AuthResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
      required:
        - token
        - newPassword
...
    VkExchangeRequest:
      type: object
      properties:
        code:
          type: string
        redirectUri:
          type: string
        state:
          type: string
      required:
        - code
        - redirectUri
        - state