# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUser } from './publicUser';

export interface AuthSession {
  token: string;
  user: PublicUser;
  mailerWarning?: string | null;
}