# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicUser {
  id: number;
  email: string | null;
  displayName: string | null;
  vkLinked: boolean;
  emailVerified: boolean;
  birthDate?: string | null;
  birthTime?: string | null;
  gender?: string | null;
}