# lib/api-zod/src/generated/types/updateMeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMeRequest {
  displayName?: string | null;
  birthDate?: string | null;
  birthTime?: string | null;
  gender?: string | null;
}