# lib/api-zod/src/generated/types/vkExchangeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VkExchangeRequest {
  code: string;
  redirectUri: string;
  state: string;
}